dbrx-simple-llm-demo/gradio_app/requirements.txt
gradio
databricks-sdk
pandas
fastapi
langchain
numpy
mlflow